pyproject.toml
[project]
name = "shrif-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery<6",
    "django>=6.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "httpx>=0.28.1",
    "langchain==0.2.11",
    "langchain-community==0.2.10",
    "langchain-core>=0.2.23",
    "langchain-openai>=0.1.25",
    "langgraph>=0.4.5",
    "langsmith>=0.1.147",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "python-telegram-bot==22.5",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "telethon>=1.42.0",
]
